# Conflicts:
#	USE_CORE/Assets/Prefabs/AudioButton.prefab
#	USE_CORE/Assets/Prefabs/AudioButton.prefab.meta
#	USE_CORE/Assets/_Scripts/M_USE/M_USE Hierarchical Finite State Machine/USE_ExperimentTemplate_Session.cs
#	USE_CORE/Assets/_Scripts/M_USE/M_USE Hierarchical Finite State Machine/USE_ExperimentTemplate_Task.cs
#	USE_CORE/Assets/_Scripts/M_USE/M_USE Hierarchical Finite State Machine/USE_ExperimentTemplate_Trial.cs
#	USE_CORE/Assets/_Scripts/M_USE/M_USE Modules/FBControllers/HaloFBController/HaloFBController.cs
#	USE_CORE/Assets/_Scripts/M_USE/M_USE Modules/StimulusHandling/USE_StimulusManagement.cs
#	USE_CORE/Assets/_USE_Session/_TaskSelection.unity
#	USE_CORE/Assets/_USE_Tasks/FeatureUncertaintyWM/FeatureUncertaintyWM_TaskLevel.cs
#	USE_CORE/Assets/_USE_Tasks/VisualSearch/VisualSearch_TrialLevel.cs
#	USE_CORE/Assets/_USE_Tasks/WWW_2D.meta

+++ USE_CORE/Assets/_USE_Tasks/THR/THR_Namespace.cs
@@ -48,9 +48,9 @@ namespace THR_Namespace
         public float MaxTouchDuration;
         public float TouchToRewardDelay;
         public float ReleaseToRewardDelay;
-        public int SquareSize;
-        public int SquareSizeMin;
-        public int SquareSizeMax;
+        public float SquareSize;
+        public float SquareSizeMin;
+        public float SquareSizeMax;
         public int PositionX;
         public int PositionX_Min;
         public int PositionX_Max;
@@ -141,9 +141,9 @@ namespace THR_Namespace
         public float MaxTouchDuration;
         public float TouchToRewardDelay;
         public float ReleaseToRewardDelay;
-        public int SquareSize;
-        public int SquareSizeMin;
-        public int SquareSizeMax;
+        public float SquareSize;
+        public float SquareSizeMin;
+        public float SquareSizeMax;
         public int PositionX;
         public int PositionX_Min;
         public int PositionX_Max;
